Safari-611.1.10/LayoutTests/css3/flexbox/nested-flexbox-min-size-auto.html

<!DOCTYPE html>
<link href="resources/flexbox.css" rel="stylesheet">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<style>
#container {
  height: 300px;
  outline: 2px solid black;
}

.inner
{
  width: 400px;
  flex: 1;
  background-color: green;
}
</style>
<script>
function change() {
  var container = document.getElementById('container');
  container.offsetHeight;
  container.style.height = '80px';
  checkLayout('#container');
}
</script>
<body onload="change()">
<p>Green rectangle should be entirely within the black rectangle</p>
<div id="log"></div>
<div id="container">
  <div class="flexbox column" style="height: 100%;">
    <div class="flexbox flex-one">
        <div class="flexbox column">
          <div class="flexbox column flex-one">
            <div class="inner" data-expected-height="80">
            </div>
          </div>
        </div>
    </div>
  </div>
</div>
